+++ myems-api/core/distributionsystem.py
@@ -651,9 +651,8 @@ class DistributionSystemClone:
             timezone_offset = int(config.utc_offset[1:3]) * 60 + int(config.utc_offset[4:6])
             if config.utc_offset[0] == '-':
                 timezone_offset = -timezone_offset
-            new_name = (str.strip(meta_result['name'])
-                        + (datetime.now()
-                           + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
+            new_name = (str.strip(meta_result['name']) +
+                        (datetime.utcnow() + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
             add_values = (" INSERT INTO tbl_distribution_systems "
                           "    (name, uuid, svg_id, description) "
                           " VALUES (%s, %s, %s, %s) ")