+++ myems-api/core/offlinemeter.py
@@ -966,9 +966,8 @@ class OfflineMeterClone:
             timezone_offset = int(config.utc_offset[1:3]) * 60 + int(config.utc_offset[4:6])
             if config.utc_offset[0] == '-':
                 timezone_offset = -timezone_offset
-            new_name = (str.strip(meta_result['name'])
-                        + (datetime.now()
-                           + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
+            new_name = (str.strip(meta_result['name']) +
+                        (datetime.utcnow() + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
             add_values = (" INSERT INTO tbl_offline_meters "
                           "    (name, uuid, energy_category_id, "
                           "     is_counted, hourly_low_limit, hourly_high_limit, "