+++ myems-api/core/photovoltaicpowerstation.py
@@ -9,7 +9,7 @@ import config
 
 class PhotovoltaicPowerStationCollection:
     def __init__(self):
-        """"Initializes PhotovoltaicPowerStationCollection"""
+        """"Initializes Class"""
         pass
 
     @staticmethod
@@ -47,6 +47,7 @@ class PhotovoltaicPowerStationCollection:
                                             "uuid": row[2]}
 
         svg_dict = dict()
+
         query = (" SELECT id, name, uuid "
                  " FROM tbl_svgs ")
         cursor.execute(query)
@@ -58,16 +59,16 @@ class PhotovoltaicPowerStationCollection:
                                     "uuid": row[2]}
 
         query = (" SELECT id, name, uuid, "
-                 "        address, postal_code, latitude, longitude, rated_power, "
-                 "        contact_id, cost_center_id, svg_id, description "
+                 "        address, postal_code, latitude, longitude, rated_capacity, rated_power, "
+                 "        contact_id, cost_center_id, svg_id, is_cost_data_displayed, phase_of_lifecycle, description "
                  " FROM tbl_photovoltaic_power_stations "
                  " ORDER BY id ")
         cursor.execute(query)
-        rows_spaces = cursor.fetchall()
+        rows_photovoltaic_power_stations = cursor.fetchall()
 
         result = list()
-        if rows_spaces is not None and len(rows_spaces) > 0:
-            for row in rows_spaces:
+        if rows_photovoltaic_power_stations is not None and len(rows_photovoltaic_power_stations) > 0:
+            for row in rows_photovoltaic_power_stations:
                 meta_result = {"id": row[0],
                                "name": row[1],
                                "uuid": row[2],
@@ -75,11 +76,14 @@ class PhotovoltaicPowerStationCollection:
                                "postal_code": row[4],
                                "latitude": row[5],
                                "longitude": row[6],
-                               "rated_power": row[7],
-                               "contact": contact_dict.get(row[8], None),
-                               "cost_center": cost_center_dict.get(row[9], None),
-                               "svg": svg_dict.get(row[10], None),
-                               "description": row[11],
+                               "rated_capacity": row[7],
+                               "rated_power": row[8],
+                               "contact": contact_dict.get(row[9], None),
+                               "cost_center": cost_center_dict.get(row[10], None),
+                               "svg": svg_dict.get(row[11], None),
+                               "is_cost_data_displayed": bool(row[12]),
+                               "phase_of_lifecycle": row[13],
+                               "description": row[14],
                                "qrcode": 'photovoltaicpowerstation:' + row[2]}
                 result.append(meta_result)
 
@@ -140,6 +144,14 @@ class PhotovoltaicPowerStationCollection:
                                    description='API.INVALID_LONGITUDE_VALUE')
         longitude = new_values['data']['longitude']
 
+        if 'rated_capacity' not in new_values['data'].keys() or \
+                not (isinstance(new_values['data']['rated_capacity'], float) or
+                     isinstance(new_values['data']['rated_capacity'], int)) or \
+                new_values['data']['rated_capacity'] <= 0.0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_RATED_CAPACITY')
+        rated_capacity = new_values['data']['rated_capacity']
+
         if 'rated_power' not in new_values['data'].keys() or \
                 not (isinstance(new_values['data']['rated_power'], float) or
                      isinstance(new_values['data']['rated_power'], int)) or \
@@ -169,6 +181,19 @@ class PhotovoltaicPowerStationCollection:
                                    description='API.INVALID_SVG_ID')
         svg_id = new_values['data']['svg_id']
 
+        if 'is_cost_data_displayed' not in new_values['data'].keys() or \
+                not isinstance(new_values['data']['is_cost_data_displayed'], bool):
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_IS_COST_DATA_DISPLAYED')
+        is_cost_data_displayed = new_values['data']['is_cost_data_displayed']
+
+        if 'phase_of_lifecycle' not in new_values['data'].keys() or \
+                not isinstance(new_values['data']['phase_of_lifecycle'], str) or \
+                len(str.strip(new_values['data']['phase_of_lifecycle'])) == 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_PHASE_OF_LIFECYCLE')
+        phase_of_lifecycle = str.strip(new_values['data']['phase_of_lifecycle'])
+
         if 'description' in new_values['data'].keys() and \
                 new_values['data']['description'] is not None and \
                 len(str(new_values['data']['description'])) > 0:
@@ -191,7 +216,7 @@ class PhotovoltaicPowerStationCollection:
         cursor.execute(" SELECT name "
                        " FROM tbl_contacts "
                        " WHERE id = %s ",
-                       (new_values['data']['contact_id'],))
+                       (contact_id,))
         row = cursor.fetchone()
         if row is None:
             cursor.close()
@@ -202,7 +227,7 @@ class PhotovoltaicPowerStationCollection:
         cursor.execute(" SELECT name "
                        " FROM tbl_cost_centers "
                        " WHERE id = %s ",
-                       (new_values['data']['cost_center_id'],))
+                       (cost_center_id,))
         row = cursor.fetchone()
         if row is None:
             cursor.close()
@@ -222,19 +247,22 @@ class PhotovoltaicPowerStationCollection:
                                    description='API.SVG_NOT_FOUND')
 
         add_values = (" INSERT INTO tbl_photovoltaic_power_stations "
-                      "    (name, uuid, address, postal_code, latitude, longitude, rated_power, "
-                      "     contact_id, cost_center_id, svg_id, description) "
-                      " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) ")
+                      " (name, uuid, address, postal_code, latitude, longitude, rated_capacity, rated_power, "
+                      "  contact_id, cost_center_id, svg_id, is_cost_data_displayed, phase_of_lifecycle, description) "
+                      " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) ")
         cursor.execute(add_values, (name,
                                     str(uuid.uuid4()),
                                     address,
                                     postal_code,
                                     latitude,
                                     longitude,
+                                    rated_capacity,
                                     rated_power,
                                     contact_id,
                                     cost_center_id,
                                     svg_id,
+                                    is_cost_data_displayed,
+                                    phase_of_lifecycle,
                                     description))
         new_id = cursor.lastrowid
         cnx.commit()
@@ -247,7 +275,7 @@ class PhotovoltaicPowerStationCollection:
 
 class PhotovoltaicPowerStationItem:
     def __init__(self):
-        """"Initializes PhotovoltaicPowerStationItem"""
+        """"Initializes Class"""
         pass
 
     @staticmethod
@@ -300,8 +328,8 @@ class PhotovoltaicPowerStationItem:
                                     "uuid": row[2]}
 
         query = (" SELECT id, name, uuid, "
-                 "        address, postal_code, latitude, longitude, rated_power, "
-                 "        contact_id, cost_center_id, svg_id, description "
+                 "        address, postal_code, latitude, longitude, rated_capacity, rated_power, "
+                 "        contact_id, cost_center_id, svg_id, is_cost_data_displayed, phase_of_lifecycle, description "
                  " FROM tbl_photovoltaic_power_stations "
                  " WHERE id = %s ")
         cursor.execute(query, (id_,))
@@ -320,11 +348,14 @@ class PhotovoltaicPowerStationItem:
                            "postal_code": row[4],
                            "latitude": row[5],
                            "longitude": row[6],
-                           "rated_power": row[7],
-                           "contact": contact_dict.get(row[8], None),
-                           "cost_center": cost_center_dict.get(row[9], None),
-                           "svg": svg_dict.get(row[10], None),
-                           "description": row[11],
+                           "rated_capacity": row[7],
+                           "rated_power": row[8],
+                           "contact": contact_dict.get(row[9], None),
+                           "cost_center": cost_center_dict.get(row[10], None),
+                           "svg": svg_dict.get(row[11], None),
+                           "is_cost_data_displayed": bool(row[12]),
+                           "phase_of_lifecycle": row[13],
+                           "description": row[14],
                            "qrcode": 'photovoltaicpowerstation:' + row[2]}
 
         resp.text = json.dumps(meta_result)
@@ -349,7 +380,8 @@ class PhotovoltaicPowerStationItem:
             raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
                                    description='API.PHOTOVOLTAIC_POWER_STATION_NOT_FOUND')
 
-        cursor.execute(" DELETE FROM tbl_photovoltaic_power_stations WHERE id = %s ", (id_,))
+        cursor.execute(" DELETE FROM tbl_photovoltaic_power_stations "
+                       " WHERE id = %s ", (id_,))
         cnx.commit()
 
         cursor.close()
@@ -414,6 +446,14 @@ class PhotovoltaicPowerStationItem:
                                    description='API.INVALID_LONGITUDE_VALUE')
         longitude = new_values['data']['longitude']
 
+        if 'rated_capacity' not in new_values['data'].keys() or \
+                not (isinstance(new_values['data']['rated_capacity'], float) or
+                     isinstance(new_values['data']['rated_capacity'], int)) or \
+                new_values['data']['rated_capacity'] <= 0.0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_RATED_CAPACITY')
+        rated_capacity = new_values['data']['rated_capacity']
+
         if 'rated_power' not in new_values['data'].keys() or \
                 not (isinstance(new_values['data']['rated_power'], float) or
                      isinstance(new_values['data']['rated_power'], int)) or \
@@ -442,6 +482,19 @@ class PhotovoltaicPowerStationItem:
             raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
                                    description='API.INVALID_SVG_ID')
         svg_id = new_values['data']['svg_id']
+
+        if 'is_cost_data_displayed' not in new_values['data'].keys() or \
+                not isinstance(new_values['data']['is_cost_data_displayed'], bool):
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_IS_COST_DATA_DISPLAYED_VALUE')
+        is_cost_data_displayed = new_values['data']['is_cost_data_displayed']
+
+        if 'phase_of_lifecycle' not in new_values['data'].keys() or \
+                not isinstance(new_values['data']['phase_of_lifecycle'], str) or \
+                len(str.strip(new_values['data']['phase_of_lifecycle'])) == 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_PHASE_OF_LIFECYCLE')
+        phase_of_lifecycle = str.strip(new_values['data']['phase_of_lifecycle'])
 
         if 'description' in new_values['data'].keys() and \
                 new_values['data']['description'] is not None and \
@@ -505,18 +558,23 @@ class PhotovoltaicPowerStationItem:
                                    description='API.SVG_NOT_FOUND')
 
         update_row = (" UPDATE tbl_photovoltaic_power_stations "
-                      " SET name = %s, address = %s, postal_code = %s, latitude = %s, longitude = %s, rated_power = %s,"
-                      "     contact_id = %s, cost_center_id = %s, svg_id = %s, description = %s "
+                      " SET name = %s, address = %s, postal_code = %s, latitude = %s, longitude = %s, "
+                      "     rated_capacity = %s, rated_power = %s, "
+                      "     contact_id = %s, cost_center_id = %s, "
+                      "     svg_id = %s, is_cost_data_displayed = %s, phase_of_lifecycle = %s, description = %s "
                       " WHERE id = %s ")
         cursor.execute(update_row, (name,
                                     address,
                                     postal_code,
                                     latitude,
                                     longitude,
+                                    rated_capacity,
                                     rated_power,
                                     contact_id,
                                     cost_center_id,
                                     svg_id,
+                                    is_cost_data_displayed,
+                                    phase_of_lifecycle,
                                     description,
                                     id_))
         cnx.commit()
@@ -529,7 +587,7 @@ class PhotovoltaicPowerStationItem:
 
 class PhotovoltaicPowerStationExport:
     def __init__(self):
-        """"Initializes PhotovoltaicPowerStationExport"""
+        """"Initializes Class"""
         pass
 
     @staticmethod
@@ -583,8 +641,8 @@ class PhotovoltaicPowerStationExport:
                                     "uuid": row[2]}
 
         query = (" SELECT id, name, uuid, "
-                 "        address, postal_code, latitude, longitude, rated_power, "
-                 "        contact_id, cost_center_id, svg_id, description "
+                 "        address, postal_code, latitude, longitude, rated_capacity, rated_power, "
+                 "        contact_id, cost_center_id, svg_id, is_cost_data_displayed, phase_of_lifecycle, description "
                  " FROM tbl_photovoltaic_power_stations "
                  " WHERE id = %s ")
         cursor.execute(query, (id_,))
@@ -596,18 +654,20 @@ class PhotovoltaicPowerStationExport:
             raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
                                    description='API.PHOTOVOLTAIC_POWER_STATION_NOT_FOUND')
         else:
-            meta_result = {"id": row[0],
-                           "name": row[1],
+            meta_result = {"name": row[1],
                            "uuid": row[2],
                            "address": row[3],
                            "postal_code": row[4],
                            "latitude": row[5],
                            "longitude": row[6],
-                           "rated_power": row[7],
-                           "contact": contact_dict.get(row[8], None),
-                           "cost_center": cost_center_dict.get(row[9], None),
-                           "svg": svg_dict.get(row[10], None),
-                           "description": row[11]}
+                           "rated_capacity": row[7],
+                           "rated_power": row[8],
+                           "contact": contact_dict.get(row[9], None),
+                           "cost_center": cost_center_dict.get(row[10], None),
+                           "svg": svg_dict.get(row[11], None),
+                           "is_cost_data_displayed": bool(row[12]),
+                           "phase_of_lifecycle": row[13],
+                           "description": row[14]}
 
         resp.text = json.dumps(meta_result)
 
@@ -674,6 +734,14 @@ class PhotovoltaicPowerStationImport:
                                    description='API.INVALID_LONGITUDE_VALUE')
         longitude = new_values['longitude']
 
+        if 'rated_capacity' not in new_values.keys() or \
+                not (isinstance(new_values['rated_capacity'], float) or
+                     isinstance(new_values['rated_capacity'], int)) or \
+                new_values['rated_capacity'] <= 0.0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_RATED_CAPACITY')
+        rated_capacity = new_values['rated_capacity']
+
         if 'rated_power' not in new_values.keys() or \
                 not (isinstance(new_values['rated_power'], float) or
                      isinstance(new_values['rated_power'], int)) or \
@@ -682,14 +750,16 @@ class PhotovoltaicPowerStationImport:
                                    description='API.INVALID_RATED_POWER')
         rated_power = new_values['rated_power']
 
-        if 'id' not in new_values['contact'].keys() or \
+        if 'contact' not in new_values.keys() or \
+                'id' not in new_values['contact'].keys() or \
                 not isinstance(new_values['contact']['id'], int) or \
                 new_values['contact']['id'] <= 0:
             raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
                                    description='API.INVALID_CONTACT_ID')
         contact_id = new_values['contact']['id']
 
-        if 'id' not in new_values['cost_center'].keys() or \
+        if 'cost_center' not in new_values.keys() or \
+                'id' not in new_values['cost_center'].keys() or \
                 not isinstance(new_values['cost_center']['id'], int) or \
                 new_values['cost_center']['id'] <= 0:
             raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
@@ -703,6 +773,19 @@ class PhotovoltaicPowerStationImport:
             raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
                                    description='API.INVALID_SVG_ID')
         svg_id = new_values['svg']['id']
+
+        if 'is_cost_data_displayed' not in new_values.keys() or \
+                not isinstance(new_values['is_cost_data_displayed'], bool):
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_IS_COST_DATA_DISPLAYED')
+        is_cost_data_displayed = new_values['is_cost_data_displayed']
+
+        if 'phase_of_lifecycle' not in new_values.keys() or \
+                not isinstance(new_values['phase_of_lifecycle'], str) or \
+                len(str.strip(new_values['phase_of_lifecycle'])) == 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_PHASE_OF_LIFECYCLE')
+        phase_of_lifecycle = str.strip(new_values['phase_of_lifecycle'])
 
         if 'description' in new_values.keys() and \
                 new_values['description'] is not None and \
@@ -757,19 +840,23 @@ class PhotovoltaicPowerStationImport:
                                    description='API.SVG_NOT_FOUND')
 
         add_values = (" INSERT INTO tbl_photovoltaic_power_stations "
-                      "    (name, uuid, address, postal_code, latitude, longitude, rated_power, "
-                      "     contact_id, cost_center_id, svg_id, description) "
-                      " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) ")
+                      "    (name, uuid, address, postal_code, latitude, longitude, rated_capacity, rated_power, "
+                      "     contact_id, cost_center_id, svg_id, is_cost_data_displayed, phase_of_lifecycle,"
+                      "     description) "
+                      " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) ")
         cursor.execute(add_values, (name,
                                     str(uuid.uuid4()),
                                     address,
                                     postal_code,
                                     latitude,
                                     longitude,
+                                    rated_capacity,
                                     rated_power,
                                     contact_id,
                                     cost_center_id,
                                     svg_id,
+                                    is_cost_data_displayed,
+                                    phase_of_lifecycle,
                                     description))
         new_id = cursor.lastrowid
         cnx.commit()
@@ -801,49 +888,58 @@ class PhotovoltaicPowerStationClone:
         cursor = cnx.cursor()
 
         query = (" SELECT id, name, uuid, "
-                 "        address, postal_code, latitude, longitude, rated_power, "
-                 "        contact_id, cost_center_id, svg_id, description "
+                 "        address, postal_code, latitude, longitude, rated_capacity, rated_power, "
+                 "        contact_id, cost_center_id, svg_id, is_cost_data_displayed, phase_of_lifecycle, description "
                  " FROM tbl_photovoltaic_power_stations "
                  " WHERE id = %s ")
         cursor.execute(query, (id_,))
         row = cursor.fetchone()
 
         if row is None:
+            cursor.close()
+            cnx.close()
             raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
                                    description='API.PHOTOVOLTAIC_POWER_STATION_NOT_FOUND')
         else:
-            meta_result = {"id": row[0],
-                           "name": row[1],
+            meta_result = {"name": row[1],
                            "uuid": row[2],
                            "address": row[3],
                            "postal_code": row[4],
                            "latitude": row[5],
                            "longitude": row[6],
-                           "rated_power": row[7],
-                           "contact_id": row[8],
-                           "cost_center_id": row[9],
-                           "svg_id": row[10],
-                           "description": row[11]}
+                           "rated_capacity": row[7],
+                           "rated_power": row[8],
+                           "contact_id": row[9],
+                           "cost_center_id": row[10],
+                           "svg_id": row[11],
+                           "is_cost_data_displayed": row[12],
+                           "phase_of_lifecycle": row[13],
+                           "description": row[14]}
+
             timezone_offset = int(config.utc_offset[1:3]) * 60 + int(config.utc_offset[4:6])
             if config.utc_offset[0] == '-':
                 timezone_offset = -timezone_offset
-            new_name = (str.strip(meta_result['name'])
-                        + (datetime.now()
-                           + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
+            new_name = str.strip(meta_result['name']) + \
+                (datetime.utcnow() + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds')
+
             add_values = (" INSERT INTO tbl_photovoltaic_power_stations "
-                          "    (name, uuid, address, postal_code, latitude, longitude, rated_power, "
-                          "     contact_id, cost_center_id, svg_id, description) "
-                          " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) ")
+                          "    (name, uuid, address, postal_code, latitude, longitude, rated_capacity, rated_power, "
+                          "     contact_id, cost_center_id, svg_id, is_cost_data_displayed, phase_of_lifecycle, "
+                          "     description) "
+                          " VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) ")
             cursor.execute(add_values, (new_name,
                                         str(uuid.uuid4()),
                                         meta_result['address'],
                                         meta_result['postal_code'],
                                         meta_result['latitude'],
                                         meta_result['longitude'],
+                                        meta_result['rated_capacity'],
                                         meta_result['rated_power'],
                                         meta_result['contact_id'],
                                         meta_result['cost_center_id'],
                                         meta_result['svg_id'],
+                                        meta_result['is_cost_data_displayed'],
+                                        meta_result['phase_of_lifecycle'],
                                         meta_result['description']))
             new_id = cursor.lastrowid
             cnx.commit()