+++ myems-api/core/rule.py
@@ -748,9 +748,8 @@ class RuleClone:
         timezone_offset = int(config.utc_offset[1:3]) * 60 + int(config.utc_offset[4:6])
         if config.utc_offset[0] == '-':
             timezone_offset = -timezone_offset
-        new_name = (str.strip(result['name'])
-                    + (datetime.now()
-                       + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
+        new_name = (str.strip(result['name']) +
+                    (datetime.utcnow() + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
         add_row = (" INSERT INTO tbl_rules "
                    "             (name, uuid, category, fdd_code, priority, "
                    "              channel, expression, message_template, is_enabled, "