+++ myems-api/core/store.py
@@ -2630,9 +2630,8 @@ class StoreClone:
             timezone_offset = int(config.utc_offset[1:3]) * 60 + int(config.utc_offset[4:6])
             if config.utc_offset[0] == '-':
                 timezone_offset = -timezone_offset
-            new_name = (str.strip(meta_result['name'])
-                        + (datetime.now()
-                           + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
+            new_name = (str.strip(meta_result['name']) +
+                        (datetime.utcnow() + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
             add_values = (" INSERT INTO tbl_stores "
                           "    (name, uuid, address, latitude, longitude, area, store_type_id, "
                           "     is_input_counted, "