+++ myems-api/core/virtualmeter.py
@@ -1356,9 +1356,8 @@ class VirtualMeterClone:
         timezone_offset = int(config.utc_offset[1:3]) * 60 + int(config.utc_offset[4:6])
         if config.utc_offset[0] == '-':
             timezone_offset = -timezone_offset
-        new_name = (str.strip(meta_result['name'])
-                    + (datetime.now()
-                       + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
+        new_name = (str.strip(meta_result['name']) +
+                    (datetime.utcnow() + timedelta(minutes=timezone_offset)).isoformat(sep='-', timespec='seconds'))
         add_values = (" INSERT INTO tbl_virtual_meters "
                       "     (name, uuid, equation, energy_category_id, is_counted, "
                       "      cost_center_id, energy_item_id, description) "